Add authorization upon add / delete / update of ratings

+++ backend/MovieRadar.WebAPI/Controllers/RatingController.cs
@@ -2,6 +2,8 @@
 using Microsoft.AspNetCore.Mvc;
 using MovieRadar.Application.Interfaces;
 using MovieRadar.Domain.Entities;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
 
 namespace MovieRadar.WebAPI.Controllers
 {
@@ -36,8 +38,28 @@ namespace MovieRadar.WebAPI.Controllers
         [HttpPost]
         public async Task<ActionResult> AddRating([FromBody] Rating newRating)
         {
-            var id = await ratingService.Add(newRating);
-            return CreatedAtAction(nameof(GetRatingById), new { id }, newRating);
+            if (!int.TryParse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value, out var userId))
+            {
+                Console.WriteLine("\n\n\n" + userId);
+                return Unauthorized();
+            }
+
+            if (userId != newRating.UserId)
+                return Forbid();
+
+            try
+            {
+                var id = await ratingService.Add(newRating);
+                return CreatedAtAction(nameof(GetRatingById), new { id }, newRating);
+            }
+            catch (ArgumentException ex)
+            {
+                return BadRequest(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, $"Error updating new rating: , {ex.Message}, inner: , {ex.InnerException}");
+            }
         }
 
         [Authorize]
@@ -45,7 +67,18 @@ namespace MovieRadar.WebAPI.Controllers
         public async Task<IActionResult> UpdateRating([FromBody] Rating updatedRating, int id)
         {
             if(id != updatedRating.Id)
-                return BadRequest("Not matching id");
+                return BadRequest("Not matching rating ID");
+
+            if (!int.TryParse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value, out var userId))
+                return Unauthorized();
+
+            var ratingToUpdate = await ratingService.GetById(id);
+
+            if (ratingToUpdate == null)
+                return NotFound();
+
+            if (userId != ratingToUpdate.UserId)
+                return Forbid();
 
             try
             {
@@ -66,6 +99,17 @@ namespace MovieRadar.WebAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteRating(int id)
         {
+            if (!int.TryParse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value, out var userId))
+                return Unauthorized();
+
+            var ratingToDelete = await ratingService.GetById(id);
+
+            if (ratingToDelete == null)
+                return NotFound();
+
+            if (userId != ratingToDelete.UserId)
+                return Forbid();
+
             var deleted = await ratingService.DeleteById(id);
             return deleted ? NoContent() : NotFound();
         }
